refactor(agents): split acp-spawn helpers and drop max-lines suppression
